music_management/requirements/requirements.txt
Flask==1.0.2
SQLAlchemy==1.2.16
Flask-SQLAlchemy==2.3.2
marshmallow==2.18.0
Flask-Alembic==2.0.1